uploading the latest build
this should be exactly the same app pretty much as the current iOS one, I just don't think the permissions requests that are all required to make it work on iOS were in included with this build. I should do that merge soon, where I make the 'iosWokring' or whatver be the source of the main branch, and then re-perfect the android stuff on that, if it isnt perfect already.